Moving base classes around to separate Qual and Profile auto tuner
